migrated testQueryNoNet

+++ pyvo/dal/tests/test_QueryNoNet.py
@@ -4,7 +4,7 @@ Tests for pyvo.dal.query
 """
 from __future__ import print_function, division
 
-import os, sys, shutil, re, imp, glob
+import os, sys, shutil, re, imp, glob, tempfile
 import unittest, pdb
 from urllib2 import URLError, HTTPError
 
@@ -13,20 +13,29 @@ import pyvo.dal.dbapi2 as daldbapi
 # from astropy.io.vo import parse as votableparse
 from astropy.io.votable.tree import VOTableFile
 from pyvo.dal.query import _votableparse as votableparse
+from astropy.utils.data import get_pkg_data_filename
+from . import aTestSIAServer as testserve
 
-testdir = os.path.dirname(sys.argv[0])
-if not testdir:  testdir = "tests"
-siaresultfile = "neat-sia.xml"
-ssaresultfile = "jhu-ssa.xml"
-testserverport = 8081
+siaresultfile = "data/neat-sia.xml"
+ssaresultfile = "data/jhu-ssa.xml"
+testserverport = 8084
+testserverport += 1
 
-try:
-    t = "aTestSIAServer"
-    mod = imp.find_module(t, [testdir])
-    testserver = imp.load_module(t, mod[0], mod[1], mod[2])
-    testserver.testdir = testdir
-except ImportError as e:
-    sys.stderr.write("Can't find test server: aTestSIAServer.py:"+str(e))
+testserver = None
+
+# def setup_module(module):
+#     """
+#     module level setup: start test server
+#     """
+#     testserver = testserve.TestServer(testserverport)
+#     testserver.start()
+
+# def teardown_module(module):
+#     """
+#     shutdown the test server
+#     """
+#     if testserver and testserver.isAlive():
+#         testserver.shutdown()
 
 class DALAccessErrorTest(unittest.TestCase):
 
@@ -165,7 +174,7 @@ class DALQueryErrorTest(unittest.TestCase):
 class DALResultsTest(unittest.TestCase):
 
     def setUp(self):
-        resultfile = os.path.join(testdir, siaresultfile)
+        resultfile = get_pkg_data_filename(siaresultfile)
         self.tbl = votableparse(resultfile)
 
     def testCtor(self):
@@ -264,7 +273,7 @@ class DALResultsTest(unittest.TestCase):
 class RecordTest(unittest.TestCase):
 
     def setUp(self):
-        resultfile = os.path.join(testdir, siaresultfile)
+        resultfile = get_pkg_data_filename(siaresultfile)
         self.tbl = votableparse(resultfile)
         self.result = dalq.DALResults(self.tbl)
         self.rec = self.result.getrecord(0)
@@ -480,17 +489,19 @@ class DALQueryTest(unittest.TestCase):
 
 class QueryExecuteTest(unittest.TestCase):
 
-    def setUp(self):
-        pass
-        # self.srvr = testserver.TestServer(testserverport)
-        # self.srvr.start()
+    srvr = None
 
-    def tearDown(self):
-        pass 
-        #if self.srvr.isAlive():
-        #    self.srvr.shutdown()
-        #if self.srvr.isAlive():
-        #    print("prob")
+    @classmethod
+    def setup_class(cls):
+        cls.srvr = testserve.TestServer(testserverport)
+        cls.srvr.start()
+
+    @classmethod
+    def teardown_class(cls):
+        if cls.srvr.isAlive():
+            cls.srvr.shutdown()
+        if cls.srvr.isAlive():
+            print("prob")
 
     def testExecute(self):
         q = dalq.DALQuery("http://localhost:{0}/sia".format(testserverport))
@@ -595,7 +606,7 @@ class QueryExecuteTest(unittest.TestCase):
 class CursorTest(unittest.TestCase):
 
     def setUp(self):
-        resultfile = os.path.join(testdir, ssaresultfile)
+        resultfile = get_pkg_data_filename(ssaresultfile)
         self.tbl = votableparse(resultfile)
 
     def testCtor(self):
@@ -667,18 +678,22 @@ class DatasetNameTest(unittest.TestCase):
     base = "testds"
 
     def setUp(self):
-        resultfile = os.path.join(testdir, siaresultfile)
+        resultfile = get_pkg_data_filename(siaresultfile)
         self.tbl = votableparse(resultfile)
         self.result = dalq.DALResults(self.tbl)
         self.rec = self.result.getrecord(0)
 
-        self.cleanfiles()
+        self.outdir = tempfile.mkdtemp()
 
     def tearDown(self):
-        self.cleanfiles()
+        shutil.rmtree(self.outdir)
 
-    def cleanfiles(self):
-        files = glob.glob(os.path.join(testdir, self.base+"*.*"))
+    def cleanfiles(self, tmpdir=None):
+        if not tmpdir:
+            tmpdir = self.outdir
+        if not os.path.isdir(tmpdir):
+            return
+        files = glob.glob(os.path.join(tmpdir, self.base+"*.*"))
         for f in files:
             os.remove(f)
 
@@ -703,6 +718,7 @@ class DatasetNameTest(unittest.TestCase):
         self.assertEquals("DAT", self.rec.suggest_extension("DAT"))
 
     def testMakeDatasetName(self):
+        self.assertTrue(os.path.isdir(self.outdir))
         self.assertEquals("./dataset.dat", self.rec.make_dataset_filename())
         self.assertEquals("./goober.dat", 
                           self.rec.make_dataset_filename(base="goober"))
@@ -712,44 +728,44 @@ class DatasetNameTest(unittest.TestCase):
                           self.rec.make_dataset_filename(base="goober", 
                                                          ext="fits"))
                           
-        self.assertEquals(testdir+"/dataset.dat", 
-                          self.rec.make_dataset_filename(testdir))
+        self.assertEquals(self.outdir+"/dataset.dat", 
+                          self.rec.make_dataset_filename(self.outdir))
 
-        path = os.path.join(testdir,self.base+".dat")
+        path = os.path.join(self.outdir,self.base+".dat")
         self.assertFalse(os.path.exists(path))
         self.assertEquals(path, 
-                          self.rec.make_dataset_filename(testdir, self.base))
+                          self.rec.make_dataset_filename(self.outdir, self.base))
         open(path,'w').close()
         self.assertTrue(os.path.exists(path))
-        path = os.path.join(testdir,self.base+"-1.dat")
+        path = os.path.join(self.outdir,self.base+"-1.dat")
         self.assertEquals(path, 
-                          self.rec.make_dataset_filename(testdir, self.base))
+                          self.rec.make_dataset_filename(self.outdir, self.base))
         open(path,'w').close()
         self.assertTrue(os.path.exists(path))
-        path = os.path.join(testdir,self.base+"-2.dat")
+        path = os.path.join(self.outdir,self.base+"-2.dat")
         self.assertEquals(path, 
-                          self.rec.make_dataset_filename(testdir, self.base))
+                          self.rec.make_dataset_filename(self.outdir, self.base))
         open(path,'w').close()
         self.assertTrue(os.path.exists(path))
-        path = os.path.join(testdir,self.base+"-3.dat")
+        path = os.path.join(self.outdir,self.base+"-3.dat")
         self.assertEquals(path, 
-                          self.rec.make_dataset_filename(testdir, self.base))
+                          self.rec.make_dataset_filename(self.outdir, self.base))
                          
         self.cleanfiles()
-        open(os.path.join(testdir,self.base+".dat"),'w').close()
-        path = os.path.join(testdir,self.base+"-1.dat")
+        open(os.path.join(self.outdir,self.base+".dat"),'w').close()
+        path = os.path.join(self.outdir,self.base+"-1.dat")
         self.assertEquals(path, 
-                          self.rec.make_dataset_filename(testdir, self.base))
-        open(os.path.join(testdir,self.base+"-1.dat"),'w').close()
-        open(os.path.join(testdir,self.base+"-2.dat"),'w').close()
-        open(os.path.join(testdir,self.base+"-3.dat"),'w').close()
-        path = os.path.join(testdir,self.base+"-4.dat")
+                          self.rec.make_dataset_filename(self.outdir, self.base))
+        open(os.path.join(self.outdir,self.base+"-1.dat"),'w').close()
+        open(os.path.join(self.outdir,self.base+"-2.dat"),'w').close()
+        open(os.path.join(self.outdir,self.base+"-3.dat"),'w').close()
+        path = os.path.join(self.outdir,self.base+"-4.dat")
         self.assertEquals(path, 
-                          self.rec.make_dataset_filename(testdir, self.base))
+                          self.rec.make_dataset_filename(self.outdir, self.base))
 
         self.cleanfiles()
-        self.assertEquals(os.path.join(testdir,self.base+".dat"),
-                          self.rec.make_dataset_filename(testdir, self.base))
+        self.assertEquals(os.path.join(self.outdir,self.base+".dat"),
+                          self.rec.make_dataset_filename(self.outdir, self.base))
 
 
 __all__ = "DALAccessErrorTest DALServiceErrorTest DALQueryErrorTest RecordTest EnsureBaseURLTest DALServiceTest DALQueryTest QueryExecuteTest CursorTest DatasetNameTest".split()
@@ -760,7 +776,16 @@ def suite():
     return unittest.TestSuite(tests)
 
 if __name__ == "__main__":
-    srvr = testserver.TestServer(testserverport)
+    try:
+        module = find_current_module(1, True)
+        pkgdir = os.path.dirname(module.__file__)
+        t = "aTestSIAServer"
+        mod = imp.find_module(t, [pkgdir])
+        testserve = imp.load_module(t, mod[0], mod[1], mod[2])
+    except ImportError as e:
+        sys.stderr.write("Can't find test server: aTestSIAServer.py:"+str(e))
+
+    srvr = testserve.TestServer(testserverport)
     try:
         srvr.start()
         unittest.main()
